+++ CollectionHub/Data/Model/Transaction.cs
@@ -6,13 +6,17 @@ namespace CollectionHub.Data.Model
     public class Transaction
     {
         [Key]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         // =========================
         // VENDEDOR
         // =========================
 
-        public long SellerId { get; set; }
+        /// <summary>
+        /// FK para o vendedor (MyUser)
+        /// </summary>
+
+        public int SellerId { get; set; }
 
         [ForeignKey(nameof(SellerId))]
         public MyUser? Seller { get; set; }
@@ -21,7 +25,11 @@ namespace CollectionHub.Data.Model
         // COMPRADOR
         // =========================
 
-        public long BuyerId { get; set; }
+
+        /// <summary>
+        /// FK para o comprador (MyUser)
+        /// </summary>
+        public int BuyerId { get; set; }
 
         [ForeignKey(nameof(BuyerId))]
         public MyUser? Buyer { get; set; }
@@ -30,12 +38,20 @@ namespace CollectionHub.Data.Model
         // ITEM
         // =========================
 
-        public long ItemId { get; set; }
+        /// <summary>
+        /// FK para o item (Item)
+        /// </summary>
+
+        public int ItemId { get; set; }
 
         [ForeignKey(nameof(ItemId))]
         public Item? Item { get; set; }
 
         // =========================
+
+        /// <summary>
+        /// Data em que a transação ocorreu
+        /// </summary>
 
         public DateTime Date { get; set; } = DateTime.Now;
