+++ CollectionHub/Data/Model/UserItem.cs
@@ -4,12 +4,12 @@ namespace CollectionHub.Data.Model
 {
     public class UserItem
     {
-        public long UserId { get; set; }
+        public int UserId { get; set; }
 
         [ForeignKey(nameof(UserId))]
         public MyUser? User { get; set; }
 
-        public long ItemId { get; set; }
+        public int ItemId { get; set; }
 
         [ForeignKey(nameof(ItemId))]
         public Item? Item { get; set; }
